🐛 fix(users): manager 파일 위치 변경
